[ALPHA-10] feat: change `get-streams-by-service-ids` handler, now it …

+++ rust-backend/src/sniffer/external_types.rs
@@ -3,5 +3,5 @@ use std::collections::HashMap;
 use tokio::sync::Mutex;
 
 lazy_static! {
-    pub static ref PORTS_TO_SNIFF: Mutex<HashMap<i16, ()>> = Mutex::new(HashMap::new());
+    pub static ref PORTS_TO_SNIFF: Mutex<HashMap<i32, ()>> = Mutex::new(HashMap::new());
 }
